Fixes negative tn values in opti
Caused by double instead of unsigned long long

+++ source/commands/sensspeccommand.h
@@ -47,7 +47,7 @@ private:
 	vector<string> outputNames;
 	set<string> labels; //holds labels to be used
 
-	long int truePositives, falsePositives, trueNegatives, falseNegatives;
+	unsigned long long truePositives, falsePositives, trueNegatives, falseNegatives;
 	bool abort, allLines, square;
 	double cutoff;
 	int precision;
